refactors ai and push modules to the new lib module
