+++ src/server/lib/push.ts
@@ -1,6 +1,7 @@
 import webPush, { PushSubscription } from "web-push";
 import { elasticsearchClient, index, getNotifications } from "server";
-import { domainName } from ".";
+
+export const domainName = process.env.DOMAIN || "mydomain";
 
 const { PUSH_VAPID_PUBLIC_KEY, PUSH_VAPID_PRIVATE_KEY } = process.env;
