+++ src/server/routes/lib/index.ts
@@ -2,6 +2,3 @@ export * from "./mails";
 export { default as Mail } from "./mails";
 export * from "./users";
 export { default as User } from "./users";
-export * from "./ai";
-export { default as Openai } from "./ai";
-export * from "./push";
